# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TradeXV2 API",
    "description": "Quantitative Trading Platform API - Market Data, Analytics, Scanner, Strategy, Replay, Backtesting",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/backtest/run": {
      "post": {
        "tags": [
          "Backtest"
        ],
        "summary": "Run Backtest",
        "description": "Run a backtest with the specified parameters.\n\nExecutes the backtest engine in **PURE_SIM** research mode against\nhistorical data from the Parquet data lake. Results are labeled\n``research_only`` and are not a live-execution parity guarantee.",
        "operationId": "run_backtest_api_v1_backtest_run_post",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BacktestRunRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestResultResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/backtest/results/{backtest_id}": {
      "get": {
        "tags": [
          "Backtest"
        ],
        "summary": "Get Backtest Result",
        "description": "Get backtest results for a completed run.",
        "operationId": "get_backtest_result_api_v1_backtest_results__backtest_id__get",
        "parameters": [
          {
            "name": "backtest_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Backtest Id"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestResultResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/backtest/comparison/{run_id}": {
      "get": {
        "tags": [
          "Backtest"
        ],
        "summary": "Compare Backtests",
        "description": "Compare multiple backtest runs side by side.",
        "operationId": "compare_backtests_api_v1_backtest_comparison__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "run_ids",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma-separated run IDs to compare",
              "title": "Run Ids"
            },
            "description": "Comma-separated run IDs to compare"
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Compare Backtests Api V1 Backtest Comparison  Run Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/flags": {
      "get": {
        "tags": [
          "Feature Flags"
        ],
        "summary": "List all feature flags",
        "description": "List all feature flags with their current state and configuration.",
        "operationId": "list_flags_api_v1_flags_get",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AllFlagsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/flags/{name}": {
      "get": {
        "tags": [
          "Feature Flags"
        ],
        "summary": "Get flag info",
        "description": "Get detailed information about a specific feature flag.\n\nArgs:\n    name: Flag name (e.g., \"SMART_ROUTING\").",
        "operationId": "get_flag_api_v1_flags__name__get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlagInfoResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/flags/{name}/toggle": {
      "post": {
        "tags": [
          "Feature Flags"
        ],
        "summary": "Toggle flag",
        "description": "Toggle a feature flag on/off (admin only).\n\nArgs:\n    name: Flag name (e.g., \"SMART_ROUTING\").",
        "operationId": "toggle_flag_api_v1_flags__name__toggle_post",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlagToggleResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/flags/{name}/rollout": {
      "post": {
        "tags": [
          "Feature Flags"
        ],
        "summary": "Set rollout percentage",
        "description": "Set the rollout percentage for a feature flag (admin only).\n\nArgs:\n    name: Flag name (e.g., \"SMART_ROUTING\").\n    body: Request body with percentage (0-100).",
        "operationId": "set_rollout_api_v1_flags__name__rollout_post",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RolloutUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RolloutUpdateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AllFlagsResponse": {
        "properties": {
          "flags": {
            "additionalProperties": {
              "$ref": "#/components/schemas/FlagInfoResponse"
            },
            "type": "object",
            "title": "Flags"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "flags",
          "count"
        ],
        "title": "AllFlagsResponse",
        "description": "Response for all flags."
      },
      "BacktestMetrics": {
        "properties": {
          "total_return_pct": {
            "type": "number",
            "title": "Total Return Pct"
          },
          "annualized_return_pct": {
            "type": "number",
            "title": "Annualized Return Pct"
          },
          "sharpe_ratio": {
            "type": "number",
            "title": "Sharpe Ratio"
          },
          "sortino_ratio": {
            "type": "number",
            "title": "Sortino Ratio"
          },
          "max_drawdown_pct": {
            "type": "number",
            "title": "Max Drawdown Pct"
          },
          "profit_factor": {
            "type": "number",
            "title": "Profit Factor"
          },
          "win_rate": {
            "type": "number",
            "title": "Win Rate"
          },
          "total_trades": {
            "type": "integer",
            "title": "Total Trades"
          },
          "winning_trades": {
            "type": "integer",
            "title": "Winning Trades"
          },
          "losing_trades": {
            "type": "integer",
            "title": "Losing Trades"
          }
        },
        "type": "object",
        "required": [
          "total_return_pct",
          "annualized_return_pct",
          "sharpe_ratio",
          "sortino_ratio",
          "max_drawdown_pct",
          "profit_factor",
          "win_rate",
          "total_trades",
          "winning_trades",
          "losing_trades"
        ],
        "title": "BacktestMetrics",
        "description": "Backtest performance metrics.\n\nCanonical definition lives in ``domain.backtest.models``.\nRe-exported here for backward compatibility."
      },
      "BacktestResultResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "timeframe": {
            "type": "string",
            "title": "Timeframe"
          },
          "metrics": {
            "$ref": "#/components/schemas/BacktestMetrics"
          },
          "trades": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trades"
          },
          "research_mode": {
            "type": "string",
            "title": "Research Mode",
            "default": "pure_sim"
          },
          "research_only": {
            "type": "boolean",
            "title": "Research Only",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "symbol",
          "timeframe",
          "metrics"
        ],
        "title": "BacktestResultResponse",
        "description": "Backtest result.\n\nCanonical definition lives in ``domain.backtest.models``.\nRe-exported here for backward compatibility."
      },
      "BacktestRunRequest": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "years": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Years",
            "default": 1
          },
          "timeframe": {
            "type": "string",
            "title": "Timeframe",
            "default": "1d"
          },
          "initial_capital": {
            "type": "number",
            "title": "Initial Capital",
            "default": 100000
          },
          "strategy": {
            "type": "string",
            "title": "Strategy",
            "description": "Strategy name"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "strategy"
        ],
        "title": "BacktestRunRequest",
        "description": "Backtest execution request."
      },
...
      "FlagInfoResponse": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "rollout_percentage": {
            "type": "integer",
            "title": "Rollout Percentage"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "default": {
            "type": "boolean",
            "title": "Default"
          }
        },
        "type": "object",
        "required": [
          "name",
          "enabled",
          "rollout_percentage",
          "description",
          "default"
        ],
        "title": "FlagInfoResponse",
        "description": "Response for flag info."
      },
      "FlagToggleResponse": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "name",
          "enabled",
          "message"
        ],
        "title": "FlagToggleResponse",
        "description": "Response for flag toggle operation."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RolloutUpdateRequest": {
        "properties": {
          "percentage": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Percentage",
            "description": "Rollout percentage (0-100)"
          }
        },
        "type": "object",
        "required": [
          "percentage"
        ],
        "title": "RolloutUpdateRequest",
        "description": "Request to update rollout percentage."
      },
      "RolloutUpdateResponse": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "rollout_percentage": {
            "type": "integer",
            "title": "Rollout Percentage"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "name",
          "rollout_percentage",
          "message"
        ],
        "title": "RolloutUpdateResponse",
        "description": "Response for rollout update operation."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}